Cpmponents: Extract the navigable toolbar component
